backend/requirements.txt
fastapi
uvicorn
python-socketio
transformers
torch
keybert
reportlab